streamlit_ui/requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0
requests>=2.31.0

# Data Processing and Visualization
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0

# UI Enhancement
streamlit-option-menu>=0.3.6
streamlit-lottie>=0.0.5
streamlit-extras>=0.3.0

# Date and Time Handling
python-dateutil>=2.8.2

# JSON and Data Validation
pydantic>=2.0.0

# HTTP Client with Better Error Handling
httpx>=0.24.0

# Authentication and Security
bcrypt>=4.0.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Styling and Icons
streamlit-elements>=0.1.0

python-dotenv>=1.0.0
# Optional: For future enhancements
# streamlit-authenticator>=0.2.3  # For user authentication
# streamlit-chat>=0.1.1           # For chat-like interface
# streamlit-aggrid>=0.3.4         # For advanced data tables